requirements.txt
# Core - FastAPI & Web
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
jinja2>=3.1.4

# Agent Frameworks - LangGraph/LangChain only (ADK/Strands mocked)
langgraph>=0.2.45
langchain>=0.3.7
langchain-community>=0.3.5
langchain-core>=0.3.15
langchain-google-genai>=2.0.5

# MCP - Compatible versions
mcp>=1.3.0
fastapi-mcp>=0.1.0

# Database - Compatible with langchain-community
sqlalchemy>=1.4,<2.0.36
aiosqlite>=0.20.0

# Data Models & Configuration
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Utilities
faker>=33.0.0
httpx>=0.28.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0

# Async & Logging
asyncio>=3.4.0
structlog>=24.4.0

# RAG - Core dependencies
chromadb>=0.5.23
sentence-transformers>=3.3.0
pypdf>=5.1.0
python-docx>=1.1.0
google-generativeai>=0.8.0
